// src/types/store.types.ts
export type ThemeConfigState = {
  isDarkMode: boolean
  mainLayout: string
  theme: string
  menu: string
  layout: string
  rtlClass: string
  animation: string
  navbar: string
  locale: string
  sidebar: boolean
  pageTitle: string
  languageList: Array<{ code: string; name: string }>
  semidark: boolean
}